fix: use relative path for scss import
This way the package can be included in other projects. The way it was before, it would look for the scss file in the wrong directory (starts looking at the root of ptoject that is including this package)

This way, it always knows where to find the scss file and thus this package can be imported in other packages/projects.